+++ src/main/kotlin/us/craigmiller160/tolkienai/server/web/type/PageResponse.kt
@@ -1,5 +1,5 @@
 package us.craigmiller160.tolkienai.server.web.type
 
 interface PageResponse : PageRequest {
-  val totalRecords: Int
+  val totalRecords: Long
 }
